refactor(accessflow): puts the sign-up and login screens into a stream directory
