+++ src/screens/accessFlow/SignIn/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { KeyboardAvoidingView, Platform, Text, View } from 'react-native';
 
-import GoogleButton from './components/GoogleButton';
+import GoogleButton from '../components/GoogleButton';
 import { Container, Divider, DividerText, Form, Header, Line } from './styles';
 
-import Logo from '../../assets/logo.svg';
+import Logo from '../../../assets/logo.svg';
 
 const SignIn: React.FC = () => {
   return (
