+++ src/screens/accessFlow/SignIn/styles.ts
@@ -1,7 +1,7 @@
 import styled from 'styled-components/native';
 import Constants from 'expo-constants';
-import colors from '../../styles/colors';
-import fonts from '../../styles/fonts';
+import colors from '../../../styles/colors';
+import fonts from '../../../styles/fonts';
 
 export const Container = styled.SafeAreaView`
   flex: 1;
